# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.14"
  title: DTL Service
# during dev, should point to your local machine
host: localhost:80
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - application/octet-stream #smell, this is necessary because supertest appears to send this content type for empty posts
# format of the responses to the client (Accepts)
produces:
  - application/json
  - text/plain
  - text/html
paths:
...
  /{entityName}/{entityId}:
    x-swagger-router-controller: ORMController
    get:
      description: Get specific entity
      operationId: findEntity
      parameters:
        - name: entityName
          type: string
          in: path
          required: true
          description: entity name
        - name: entityId
          type: number
          in: path
          required: true
          description: The id of the entity          
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EntityResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: update entity
      operationId: updateEntity
      parameters:
        - name: entityName
          type: string
          in: path
          required: true
          description: Indicates this request is an automated test request
        - name: entityId
          type: number
          in: path
          required: true
          description: The id of the entity
        - name: bodyParam
          in: body
          description: JSON document that has teams and their related summonerIds and metadata
          required: true
          schema:
            $ref: "#/definitions/EntityResponse"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EntityResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Delete specific entity
      operationId: deleteEntity
      parameters:
        - name: entityName
          type: string
          in: path
          required: true
          description: entity name
        - name: entityId
          type: number
          in: path
          required: true
          description: The id of the entity
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EntityResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
  /{entityName}:
    x-swagger-router-controller: ORMController
    get:
      description: Get all entities
      operationId: findAllEntitiesSearch
      parameters:
        - name: entityName
          type: string
          in: path
          required: true
          description: entity name
        - name: select
          type: string
          in: query
          required: false
          description: list of fields to return
        - name: filter
          type: string
          in: query
          required: false
          description: search experession
        - name: sort
          type: string
          in: query
          required: false
          description: list of sort fileds
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EntityResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: creates an entity
      operationId: createEntity
      parameters:
        - name: entityName
          type: string
          in: path
          required: true
          description: Indicates this request is an automated test request
        - name: bodyParam
          in: body
          description: JSON document that has teams and their related summonerIds and metadata
          required: true
          schema:
            $ref: "#/definitions/EntityResponse"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EntityResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"                     
  /findmenow/open:
    x-swagger-router-controller: ORMController
    get:
      description: Get all entities
      operationId: findAllEntitiesSearch
      parameters:
        - name: entityName
          type: string
          in: query
          required: true
          description: entity name
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EntityResponse"
        "default":
          description: Failure
          schema:
            $ref: "#/definitions/ErrorResponse"           
# specific complex operations 
...
definitions:
...
  EntityResponse:
    type: object
  # Response definitions
  ErrorResponse:
    required:
      - type
      - code
      - message
    properties:
      type:
        type: string
      code:
        type: number
      message:
        type: string
...